# Switch from FBOs to custom callbacks for drawing maps