This branch removes FBO usage from the map manager and switches to ImGui custom callbacks, which is a much faster (and memory-efficient) drawing method because a texture is not involved.

Comparison:
```
Before (FBOs):
Default (16x16; 1 layer) - 60FPS (limited by vsync)
BIG map (64x64; 1 layer) - 39FPS
BIG with two layers (64x64) - 29FPS
EVEN BIGGER map (128x128; 1 layer) - 12FPS

After (ImGui custom callbacks):
Default (16x16; 1 layer) - 60FPS (limited by vsync)
BIG map (64x64; 1 layer) - 60FPS (limited by vsync)
BIG with two layers (64x64) - 60FPS (limited by vsync)
EVEN BIGGER map (128x128; 1 layer) - 60FPS (limited by vsync)
INSANELY BIG map (1024x1024; 1 layer) - 49FPS
```